requirements.txt
awkward==2.0.8
matplotlib==3.6.3
networkx==2.8.8
numpy==1.24.1
PyYAML==6.0
qiskit_terra==0.22.4
qutip==4.7.1
scikit_learn==1.2.1
scipy==1.10.0
six==1.16.0
tensorflow==2.11.0
uproot==5.0.3
